Migrate from H2 to MySQL with JWT authentication
- Replace H2 in-memory with persistent MySQL database
- Add User entity and repository mapped to user table
- Update Task entity to match MySQL schema (taskId, content, FK user)
- Implement JWT-based authentication (login + register)
- Add Spring Security with stateless JWT filter
- Add /login and /register endpoints returning JWT tokens
- Add /task/my endpoint for user-specific tasks
- Hash passwords with bcrypt on startup
- Create login.html with sign in / sign up toggle
- Rename index.html to tasks.html with JWT auth headers
- Add user display and logout in task UI
- Update README and ER diagram

+++ src/main/java/com/example/task_manager/TaskNotFoundException.java
@@ -2,8 +2,7 @@ package com.example.task_manager;
 
 public class TaskNotFoundException extends RuntimeException {
 
-	public TaskNotFoundException(Long id){
-		super("Could not find task " + id);
-	}
-
+    public TaskNotFoundException(Integer id) {
+        super("Could not find task " + id);
+    }
 }
